+++ emr-20210320/src/Models/ListScriptsRequest.php
@@ -1,0 +1,115 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Emr\V20210320\Models;
+
+use AlibabaCloud\Tea\Model;
+
+class ListScriptsRequest extends Model
+{
+    /**
+     * @description 集群ID。
+     *
+     * @example c-b933c5aac8fe****
+     *
+     * @var string
+     */
+    public $clusterId;
+
+    /**
+     * @description 一次获取的最大记录数。取值范围：1~100。
+     *
+     * @example 10
+     *
+     * @var int
+     */
+    public $maxResults;
+
+    /**
+     * @description 标记当前开始读取的位置，置空表示从头开始。
+     *
+     * @example dd6b1b2a-5837-5237-abe4-ff0c89568980
+     *
+     * @var string
+     */
+    public $nextToken;
+
+    /**
+     * @description 区域ID。
+     *
+     * @example cn-hangzhou
+     *
+     * @var string
+     */
+    public $regionId;
+
+    /**
+     * @description 集群脚本类型。
+     *
+     * @example BOOTSTRAP
+     *
+     * @var string
+     */
+    public $scriptType;
+    protected $_name = [
+        'clusterId'  => 'ClusterId',
+        'maxResults' => 'MaxResults',
+        'nextToken'  => 'NextToken',
+        'regionId'   => 'RegionId',
+        'scriptType' => 'ScriptType',
+    ];
+
+    public function validate()
+    {
+    }
+
+    public function toMap()
+    {
+        $res = [];
+        if (null !== $this->clusterId) {
+            $res['ClusterId'] = $this->clusterId;
+        }
+        if (null !== $this->maxResults) {
+            $res['MaxResults'] = $this->maxResults;
+        }
+        if (null !== $this->nextToken) {
+            $res['NextToken'] = $this->nextToken;
+        }
+        if (null !== $this->regionId) {
+            $res['RegionId'] = $this->regionId;
+        }
+        if (null !== $this->scriptType) {
+            $res['ScriptType'] = $this->scriptType;
+        }
+
+        return $res;
+    }
+
+    /**
+     * @param array $map
+     *
+     * @return ListScriptsRequest
+     */
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['ClusterId'])) {
+            $model->clusterId = $map['ClusterId'];
+        }
+        if (isset($map['MaxResults'])) {
+            $model->maxResults = $map['MaxResults'];
+        }
+        if (isset($map['NextToken'])) {
+            $model->nextToken = $map['NextToken'];
+        }
+        if (isset($map['RegionId'])) {
+            $model->regionId = $map['RegionId'];
+        }
+        if (isset($map['ScriptType'])) {
+            $model->scriptType = $map['ScriptType'];
+        }
+
+        return $model;
+    }
+}