+++ emr-20210320/src/Models/ListScriptsResponseBody.php
@@ -4,29 +4,24 @@
 
 namespace AlibabaCloud\SDK\Emr\V20210320\Models;
 
-use AlibabaCloud\SDK\Emr\V20210320\Models\ListApmMetadataResponseBody\metadata;
+use AlibabaCloud\SDK\Emr\V20210320\Models\ListScriptsResponseBody\scripts;
 use AlibabaCloud\Tea\Model;
 
-class ListApmMetadataResponseBody extends Model
+class ListScriptsResponseBody extends Model
 {
     /**
      * @description 本次请求所返回的最大记录条数。
      *
-     * @example 20
+     * @example 10
      *
      * @var int
      */
     public $maxResults;
 
     /**
-     * @var metadata[]
-     */
-    public $metadata;
-
-    /**
      * @description 返回读取到的数据位置，空代表数据已经读取完毕。
      *
-     * @example DD6B1B2A-5837-5237-ABE4-FF0C89568980
+     * @example dd6b1b2a-5837-5237-abe4-ff0c89568982
      *
      * @var string
      */
@@ -42,6 +37,11 @@ class ListApmMetadataResponseBody extends Model
     public $requestId;
 
     /**
+     * @var scripts[]
+     */
+    public $scripts;
+
+    /**
      * @description 本次请求条件下的数据总量。
      *
      * @example 200
@@ -51,9 +51,9 @@ class ListApmMetadataResponseBody extends Model
     public $totalCount;
     protected $_name = [
         'maxResults' => 'MaxResults',
-        'metadata'   => 'Metadata',
         'nextToken'  => 'NextToken',
         'requestId'  => 'RequestId',
+        'scripts'    => 'Scripts',
         'totalCount' => 'TotalCount',
     ];
 
@@ -67,20 +67,20 @@ class ListApmMetadataResponseBody extends Model
         if (null !== $this->maxResults) {
             $res['MaxResults'] = $this->maxResults;
         }
-        if (null !== $this->metadata) {
-            $res['Metadata'] = [];
-            if (null !== $this->metadata && \is_array($this->metadata)) {
-                $n = 0;
-                foreach ($this->metadata as $item) {
-                    $res['Metadata'][$n++] = null !== $item ? $item->toMap() : $item;
-                }
-            }
-        }
         if (null !== $this->nextToken) {
             $res['NextToken'] = $this->nextToken;
         }
         if (null !== $this->requestId) {
             $res['RequestId'] = $this->requestId;
+        }
+        if (null !== $this->scripts) {
+            $res['Scripts'] = [];
+            if (null !== $this->scripts && \is_array($this->scripts)) {
+                $n = 0;
+                foreach ($this->scripts as $item) {
+                    $res['Scripts'][$n++] = null !== $item ? $item->toMap() : $item;
+                }
+            }
         }
         if (null !== $this->totalCount) {
             $res['TotalCount'] = $this->totalCount;
@@ -92,7 +92,7 @@ class ListApmMetadataResponseBody extends Model
     /**
      * @param array $map
      *
-     * @return ListApmMetadataResponseBody
+     * @return ListScriptsResponseBody
      */
     public static function fromMap($map = [])
     {
@@ -100,20 +100,20 @@ class ListApmMetadataResponseBody extends Model
         if (isset($map['MaxResults'])) {
             $model->maxResults = $map['MaxResults'];
         }
-        if (isset($map['Metadata'])) {
-            if (!empty($map['Metadata'])) {
-                $model->metadata = [];
-                $n               = 0;
-                foreach ($map['Metadata'] as $item) {
-                    $model->metadata[$n++] = null !== $item ? metadata::fromMap($item) : $item;
-                }
-            }
-        }
         if (isset($map['NextToken'])) {
             $model->nextToken = $map['NextToken'];
         }
         if (isset($map['RequestId'])) {
             $model->requestId = $map['RequestId'];
+        }
+        if (isset($map['Scripts'])) {
+            if (!empty($map['Scripts'])) {
+                $model->scripts = [];
+                $n              = 0;
+                foreach ($map['Scripts'] as $item) {
+                    $model->scripts[$n++] = null !== $item ? scripts::fromMap($item) : $item;
+                }
+            }
         }
         if (isset($map['TotalCount'])) {
             $model->totalCount = $map['TotalCount'];