+++ emr-20210320/src/Models/ScalingRule.php
@@ -34,6 +34,11 @@ class ScalingRule extends Model
     public $metricsTrigger;
 
     /**
+     * @var int
+     */
+    public $minAdjustmentValue;
+
+    /**
      * @description 规则名称。
      *
      * @example scale-out-memory
@@ -58,12 +63,13 @@ class ScalingRule extends Model
      */
     public $triggerType;
     protected $_name = [
-        'activityType'    => 'ActivityType',
-        'adjustmentValue' => 'AdjustmentValue',
-        'metricsTrigger'  => 'MetricsTrigger',
-        'ruleName'        => 'RuleName',
-        'timeTrigger'     => 'TimeTrigger',
-        'triggerType'     => 'TriggerType',
+        'activityType'       => 'ActivityType',
+        'adjustmentValue'    => 'AdjustmentValue',
+        'metricsTrigger'     => 'MetricsTrigger',
+        'minAdjustmentValue' => 'MinAdjustmentValue',
+        'ruleName'           => 'RuleName',
+        'timeTrigger'        => 'TimeTrigger',
+        'triggerType'        => 'TriggerType',
     ];
 
     public function validate()
@@ -81,6 +87,9 @@ class ScalingRule extends Model
         }
         if (null !== $this->metricsTrigger) {
             $res['MetricsTrigger'] = null !== $this->metricsTrigger ? $this->metricsTrigger->toMap() : null;
+        }
+        if (null !== $this->minAdjustmentValue) {
+            $res['MinAdjustmentValue'] = $this->minAdjustmentValue;
         }
         if (null !== $this->ruleName) {
             $res['RuleName'] = $this->ruleName;
@@ -111,6 +120,9 @@ class ScalingRule extends Model
         }
         if (isset($map['MetricsTrigger'])) {
             $model->metricsTrigger = MetricsTrigger::fromMap($map['MetricsTrigger']);
+        }
+        if (isset($map['MinAdjustmentValue'])) {
+            $model->minAdjustmentValue = $map['MinAdjustmentValue'];
         }
         if (isset($map['RuleName'])) {
             $model->ruleName = $map['RuleName'];